COMP: Rename ctkAbstractScheduler files to ctkJobScheduler
